\name{internal}

\title{Internal functions used in package `gausspred'}

\description{This function is internal. Type the function name to look at its details}

\alias{gen_QF_lmd}

\alias{comp_adjfactor}

\alias{comp_fstat}

\alias{gen_transformer}

\alias{log_sum_exp}

\alias{partition_even}

\alias{pooled_var}

\keyword{internal}
